Reorganize the Nvidia source code into multiple files
